Implement DVFS and complete the lookahead policy
Make energy consumption a double to better reflect action usage
runtime is now ceiled to the next millisecond
Potentiall this means that an edge case could increase the critical path per task by 1ms, which in reality would probably be less. This is acceptable to us as runtimes are in the thousands of milliseconds (and larger).

+++ src/main/kotlin/science/atlarge/wta/simulator/core/TaskStateMonitor.kt
@@ -23,7 +23,7 @@ class TaskStateMonitor(
         private set
     var completedTaskCount = 0
         private set
-    var energyConsumed: Long = 0
+    var energyConsumed: Double = 0.0
         private set
 
     private val reverseDependencies: Array<Array<Task>>
